# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BlessedBites API
  version: 1.0.0
  description: |
    REST API for the BlessedBites food ordering platform.

    **Authentication**: Protected endpoints require a Bearer JWT obtained via
    `POST /auth/login`. Refresh tokens are set/rotated via an `HttpOnly`
    `bb_refresh` cookie on `/auth/refresh`. Each refresh rotates the token —
    reuse of a consumed refresh token returns `401`.
servers:
  - url: /api/v1
security: []
paths:
...
  /menu/{id}:
    get:
      summary: Get a single menu item by ID
      tags: [Menu]
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          description: Menu item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/menu:
    get:
      summary: Admin – list all menu items (including inactive)
      tags: [Admin]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Full menu list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Admin – create a new menu item
      tags: [Admin]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertMenuItemRequest'
      responses:
        '201':
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/menu/{id}:
    put:
      summary: Admin – replace a menu item
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertMenuItemRequest'
      responses:
        '200':
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Admin – delete a menu item
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived access token obtained from `POST /auth/login` or
        `POST /auth/refresh`. Valid for 30 minutes.
  parameters:
    PathID:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Bad request – validation or parse error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but insufficient permissions (admin required)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: resource not found
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        category_id:
          type: integer
        image_url:
          type: string
        is_active:
          type: boolean
      example:
        id: 1
        name: Jerk Chicken
        description: Slow-cooked with signature spices
        price: 14.99
        category_id: 2
        image_url: /uploads/jerk-chicken.jpg
        is_active: true
...
    UpsertMenuItemRequest:
      type: object
      required: [name, price, category_id]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
          minimum: 0
        category_id:
          type: integer
        image_url:
          type: string
        is_active:
          type: boolean